refactor: g5endpoint, remove auth-scheme and update the request-dto to have access-token for guest user to access

+++ src/Core/Domain/PenomyAPI.Domain.RelationalDb/Repositories/Features/Generic/IG49Repository.cs
@@ -11,5 +11,5 @@ public interface IG49Repository
     Task<ArtworkMetaData> RateArtworkAsync(long userId, long artworkId, byte starRates,
         CancellationToken cancellationToken);
 
-    Task<long> GetCurrentUserRatingAsync(long userId, long artworkId, CancellationToken cancellationToken);
+    Task<byte> GetCurrentUserRatingAsync(long userId, long artworkId, CancellationToken cancellationToken);
 }
